refactor(mockagent): split acp_runtime.go into runtime + state + control
acp_runtime.go was 226 lines — past the project's 200-line soft cap.
Extracting two single-purpose files restores readability without
diluting cohesion:

  runtime_state.go    runtimeState struct + writer/awaitResponse/
                      configMu accessors (the shared scratchpad
                      every scenario reaches into).
  acp_control.go      handleControlRequest dispatch — the central
                      switch new subtypes (set_thinking_level, etc.)
                      will extend.

Sizes after split:
  acp_runtime.go      127 lines  (main loop + dispatch + protocol helpers)
  runtime_state.go     85 lines
  acp_control.go       44 lines

No behavior change; mockagent_test still passes.
