+++ runner/internal/agents/mockagent/BUILD.bazel
@@ -3,10 +3,12 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "mockagent",
     srcs = [
+        "acp_control.go",
         "acp_runtime.go",
         "emit.go",
         "env_dump.go",
         "pty_runtime.go",
+        "runtime_state.go",
         "scenario_basic.go",
         "scenario_config.go",
         "scenario_errors.go",
